+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Added implements of `TryInto<Symbol>` and `TryInto<Production>` to ease parsing.
 * Added an implementation of `FromStr` for `ProductionString`. 
+* Added an implementation of `FromStr` for `Symbol`. 
 
 ### Changed